backend/requirements.txt
Flask
flask_cors
flask-sqlalchemy
google-generativeai
python-dotenv
stripe
sqlalchemy
mysql-connector-python